+++ src/main/java/roomescape/reservation/controller/AdminReservationTimeController.java
@@ -1,0 +1,43 @@
+package roomescape.reservation.controller;
+
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.global.auth.annotation.RoleRequired;
+import roomescape.member.entity.RoleType;
+import roomescape.reservation.dto.request.ReservationTimeCreateRequest;
+import roomescape.reservation.dto.response.ReservationTimeCreateResponse;
+import roomescape.reservation.service.ReservationTimeService;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/admin/times")
+public class AdminReservationTimeController {
+
+    private final ReservationTimeService reservationTimeService;
+
+    @PostMapping
+    @RoleRequired(roleType = RoleType.ADMIN)
+    public ResponseEntity<ReservationTimeCreateResponse> createTimeByAdmin(
+            @RequestBody @Valid ReservationTimeCreateRequest request
+    ) {
+        ReservationTimeCreateResponse response = reservationTimeService.createTime(request);
+        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    }
+
+    @DeleteMapping("/{id}")
+    @RoleRequired(roleType = RoleType.ADMIN)
+    public ResponseEntity<Void> deleteTimeByAdmin(
+            @PathVariable("id") long id
+    ) {
+        reservationTimeService.deleteTime(id);
+        return ResponseEntity.noContent().build();
+    }
+}
