+++ src/main/java/roomescape/reservation/controller/ReservationTimeController.java
@@ -1,24 +1,14 @@
 package roomescape.reservation.controller;
 
-import jakarta.validation.Valid;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.global.auth.annotation.RoleRequired;
-import roomescape.member.entity.RoleType;
-import roomescape.reservation.dto.request.ReservationTimeCreateRequest;
 import roomescape.reservation.dto.response.AvailableReservationTimeResponse;
-import roomescape.reservation.dto.response.ReservationTimeCreateResponse;
 import roomescape.reservation.dto.response.ReservationTimeReadResponse;
 import roomescape.reservation.service.ReservationTimeService;
 
@@ -28,15 +18,6 @@ import roomescape.reservation.service.ReservationTimeService;
 public class ReservationTimeController {
 
     private final ReservationTimeService reservationTimeService;
-
-    @PostMapping
-    @RoleRequired(roleType = RoleType.ADMIN)
-    public ResponseEntity<ReservationTimeCreateResponse> createTime(
-            @RequestBody @Valid ReservationTimeCreateRequest request
-    ) {
-        ReservationTimeCreateResponse response = reservationTimeService.createTime(request);
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
-    }
 
     @GetMapping
     public ResponseEntity<List<ReservationTimeReadResponse>> getAllTimes() {
@@ -51,14 +32,5 @@ public class ReservationTimeController {
     ) {
         List<AvailableReservationTimeResponse> responses = reservationTimeService.getAvailableTimes(date, themeId);
         return ResponseEntity.ok().body(responses);
-    }
-
-    @DeleteMapping("/{id}")
-    @RoleRequired(roleType = RoleType.ADMIN)
-    public ResponseEntity<Void> deleteTime(
-            @PathVariable("id") long id
-    ) {
-        reservationTimeService.deleteTime(id);
-        return ResponseEntity.noContent().build();
     }
 }
